refactor: rename dynamicert to dynacert
